Add some helper functions for 2D.

+++ src/Math/float4d.h
@@ -194,14 +194,14 @@ public:
 
 	vec ToPointVec() const
 	{
-		assert1(EqualAbs(w, 1.0), w);
-		return POINT_VEC(x, y, z);
+		assert1(EqualAbs((float)w, 1.0), w);
+		return POINT_VEC((float)x, (float)y, (float)z);
 	}
 
 	vec ToDirVec() const
 	{
-		assert1(EqualAbs(w, 0.0), w);
-		return DIR_VEC(x, y, z);
+		assert1(EqualAbs((float)w, 0.0), w);
+		return DIR_VEC((float)x, (float)y, (float)z);
 	}
 
 	bool Equals(const float4d &other, float epsilon = 1e-3f) const
